+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - lightning-cli: `help <cmd>` finds man pages even if `make install` not run.
 - JSON API: `waitsendpay` now has an `erring_channel_direction` field.
 - JSON API: `listpeers` now has a `channel_direction` field in `channels`.
+- JSON API: `listchannels` now takes a `source` option to filter by node id.
 
 ### Changed